modified user class with usertype and other classes accordingly

+++ CodeFury-WFS1-Team2-BackendApp/src/main/java/com/hsbc/ecommerceapp/service/impl/AdminServiceImpl.java
@@ -1,16 +1,21 @@
 package com.hsbc.ecommerceapp.service.impl;
 
+import com.hsbc.ecommerceapp.exceptions.ProductNotFoundException;
+import com.hsbc.ecommerceapp.exceptions.SubscriptionNotFoundException;
 import com.hsbc.ecommerceapp.model.Product;
 import com.hsbc.ecommerceapp.model.Subscription;
+import com.hsbc.ecommerceapp.model.User;
 import com.hsbc.ecommerceapp.service.AdminService;
 import com.hsbc.ecommerceapp.service.ProductService;
 import com.hsbc.ecommerceapp.service.SubscriptionService;
+import com.hsbc.ecommerceapp.storage.SubscriptionStorage;
 
 import java.util.List;
 
 public class AdminServiceImpl implements AdminService {
     private ProductService productService;
     private SubscriptionService subscriptionService;
+    private SubscriptionStorage subscriptionStorage;
 
     public AdminServiceImpl(ProductService productService, SubscriptionService subscriptionService) {
         this.productService = productService;
@@ -18,30 +23,47 @@ public class AdminServiceImpl implements AdminService {
     }
 
     @Override
-    public void addProduct(Product product) {
+    public void addProduct(User user, Product product) {
+        if(!user.isAdmin())
+            throw new SecurityException("Access denied to user!");
+
         productService.addProduct(product);
     }
 
     @Override
-    public void updateProduct(Product product) {
+    public void updateProduct(User user, Product product) {
+        if(!user.isAdmin())
+            throw new SecurityException("Access denied to user!");
         productService.updateProduct(product);
     }
 
     @Override
-    public void deleteProduct(String productId) {
+    public void deleteProduct(User user, String productId) throws ProductNotFoundException {
+        if(!user.isAdmin())
+            throw new SecurityException("Access denied to user!");
         productService.deleteProduct(productId);
     }
 
     @Override
-    public void deactivateSubscription(String subscriptionId) {
-        subscriptionService.cancelSubscription(subscriptionId);
+    public void deactivateSubscription(User user, String subscriptionId) throws SubscriptionNotFoundException {
+        if(!user.isAdmin())
+            throw new SecurityException("Access denied to user!");
+        Subscription subscription = subscriptionStorage.getSubscriptionById(subscriptionId);
+        if (subscription == null)
+            throw new SubscriptionNotFoundException("Subscription not found");
+        subscription.setActive(false);
+        subscriptionStorage.updateSubscription(subscription);
     }
 
     @Override
-    public void activateSubscription(String subscriptionId) {
-        Subscription subscription = subscriptionService.getSubscriptionById(subscriptionId);
+    public void activateSubscription(User user, String subscriptionId) {
+        if (!user.isAdmin())
+            throw new SecurityException("User does not have admin privileges.");
+        Subscription subscription = subscriptionStorage.getSubscriptionById(subscriptionId);
+        if (subscription == null)
+            throw new SubscriptionNotFoundException("Subscription not found");
         subscription.setActive(true);
-        subscriptionService.updateSubscription(subscription);
+        subscriptionStorage.updateSubscription(subscription);
     }
 
     @Override
